requirements.txt
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
numpy
Pillow
scikit-learn
SQLAlchemy
streamlit
tqdm
torch
st_clickable_images